+++ ReleaseNotes.md
@@ -32,6 +32,7 @@ Renamed functions
 - zoomTo() removed. Use setZoomLevel() instead.
 - getLayerPlugin() removed. Use getLayerPlugins(id) with id parameter to fetch reference to single plugin.
 - getMapScales() renamed getScaleArray() to be consistent with getResolution()/getResolutionArray().
+- calculateScaleResolution() renamed getResolutionForScale().
 
 Unused functions removed:
   - _ensureExists()